% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ddmmss.R
\name{build_ddmmss}
\alias{build_ddmmss}
\title{Make ddmmss/dddmmss strings from numeric coordinates}
\usage{
build_ddmmss(x)
}
\arguments{
\item{x}{\code{numeric} numeric vector containing the coordinates to convert into string.}
}
\value{
A \code{character} vector of the same length as \code{x}. Strings have a minimum length of 5 characters without
the sign, not including the sign, and a maximum length of 8 characters, including the sign.
}
\description{
\code{build_ddmmss} builds coordinate strings of the type +/-ddmmss, or +/-dddmmss from coordinates in numeric form,
where dd/ddd, mm and ss stand for degrees, minutes and seconds, respectively.
}
\details{
If the input is negative, the output will also be negative. That would be equivalent to a -ddmmss/-dddmmsss case.
}
\examples{
x <- c("-25005", "2560032", "25000")
y <- sapply(x, function(z) extract_ddmmss(z))
z <- build_ddmmss(y)
}
