% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_CNIG_INE.R
\name{check_CNIG_INE}
\alias{check_CNIG_INE}
\title{Check CNIG and INE datasets}
\usage{
check_CNIG_INE(munic_CNIG, munic_INE, verbose = T)
}
\arguments{
\item{munic_CNIG}{\code{data.frame} containing the 'MUNICIPIOS' dataset from the
\emph{Nomenclator geografico de municipios y entidades de poblacion} at the CNIG web site.}

\item{munic_INE}{\code{data.frame} containing the 'diccionario24' dataset from the
INE web site.}
}
\value{
Informative messages.
}
\description{
This function checks the consistency of municipal names in the CNIG and INE databases that accompany
this package.
}
\details{
This function is not exported because it is usually used by the package maintainer only.

Municipal names have the case lowered, split by slashes into two string, quotation marks
modified and commas removed. During the split, the original name is split into 'left' and 'right'
versions (which will usually correspond to the Spanish and local language names). Those two versions
can be used for checking, although at this moment only 'left' names are used.
}
\examples{
data(munic_CNIG)
data(munic_INE)
check_CNIG_INE(munic_CNIG, munic_INE)
}
