% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_df.R
\name{compare_df}
\alias{compare_df}
\title{Compare two or more data.frames}
\usage{
compare_df(...)
}
\arguments{
\item{...}{comma-separated 'data.frames' to be compared.}
}
\value{
'TRUE' if the comparisons were successful; FALSE if they were not.
}
\description{
\code{data.frames} are checked whether their column numbers, column names
and column formats all match.
}
\details{
The content of the columns do not need to match; that is, inputs can have
different number of rows and contain different numbers/characters/anything.

Messsages are printed on screen showing the progress of the comparison:
\itemize{
\item Inputs are checked for "data.framness". If any fails a 'return(FALSE)'
is executed.
\item Column numbers must match.
\item Column names must match one by one. Obviously, if the previous conditions fails,
this will also fail.
\item Column format must match one by one.
}
}
\examples{
df1 <- as.data.frame(matrix(100, 20, 5))
colnames(df1) <- letters[1:5]
df2 <- as.data.frame(matrix(80, 20, 4))
colnames(df2) <- letters[2:5]

# Success.
compare_df(df1, df1)

# Failure.
compare_df(df1, df2)
compare_df(runif(10), df1)

}
