% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction_coordinates.R
\name{correction_coordinates}
\alias{correction_coordinates}
\title{Title}
\usage{
correction_coordinates(df)
}
\arguments{
\item{x}{}
}
\description{
Title
}
\details{
A correction to coordinates, email from Carmen Mirta Dimas Suárez (CEH) on July 17th, 2024.
}
\examples{
df <- data.frame(indroea = c("3179", "3000"), lat = 1:2, long = 3:4)
correction_coordinates(df)
}
