% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curly_quotes.R
\name{curly_quotes}
\alias{curly_quotes}
\title{Substitute curly quotes by straight ones.}
\usage{
curly_quotes(x, single = T)
}
\arguments{
\item{x}{character vector.}

\item{single}{logical, if set to TRUE (default) any double curly quotation mark
will be substituted by a single straight quotation mark. Otherwise, a double curly
quotation mark will be subsituted by a double straight quotation mark.}
}
\value{
A character vector of the same length as \code{x} with curly quotation marks
substituted.
}
\description{
\code{curly_quotes} performs a substitution of curly quotation marks (left or right,
single or double) by straight quotation marks.
}
\details{
Simple implementation of the \code{gsub} command.
}
\examples{
x <- c("Straight quote'", "Single straight '‘and right curly quotes ",
"Double “ ”and ” single ‘curly ’ ''quotes")
print(rbind(original = x, single = curly_quotes(x), double = curly_quotes(x, F)))
}
