% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_to_monthly.R
\name{daily_to_monthly}
\alias{daily_to_monthly}
\title{Convert \emph{cedex} daily data to monthly}
\usage{
daily_to_monthly(x, table = NULL)
}
\arguments{
\item{x}{\code{data.frame}}

\item{table}{\code{character} with the name of the type of table, which in turn will
determine the columns in \code{x}. Default value is "afliq".}
}
\value{
A \code{data.frame} with columns "Station", "Date", "Height_mean", "Flow_mean" and "Flow",
corresponding to the station name, date, mean monthly height and flow, and total monthly flow.
}
\description{
\code{daily_to_monthly} calculates monthly averages and totals of the "Height" and
"Flow" columns, respectively.
}
\details{
Total monthly flow for February is calculated accounting for a leap year, if required.
NA's are eliminated and monthly total values are then computed by projecting to the
number of days per month.
}
\examples{
.
# All stations at once.
x1 <- read_cedex_basin()
y1 <- daily_to_monthly(x1)

# Only one station.
x2 <- read_cedex_txt("https://ceh.cedex.es/anuarioaforos/temp/resultado8187.txt", "daily")
y2 <- daily_to_monthly(x2)

# They match.
i <- match(paste0(y2$Station, y2$Flow), paste0(y1$Station, y1$Flow))
print(sum(abs(y2$Flow - y1$Flow[i]), na.rm = T))

}
