% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_conversion.R
\name{date_conversion}
\alias{date_conversion}
\title{Title}
\usage{
date_conversion(julian_date)
}
\arguments{
\item{julian_date}{}
}
\description{
Title
}
\details{
If 'julian_date' is before ***, it is assumed to be from the Julian calendar. If not, it is
assumed to be a Gregorian date and the output is identical to the input.
}
