% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_per_month.R
\docType{data}
\name{days_per_month}
\alias{days_per_month}
\title{Number of days per month for a non-leap year}
\format{
A named numeric vector.
}
\usage{
days_per_month
}
\description{
A named numeric vector containing the number of days per month for a non-leap year.
Months labels given in English.
}
\author{
R. Molowny-Horas
}
\keyword{date}
