% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dia_abr.R
\docType{data}
\name{dia_abr}
\alias{dia_abr}
\title{Abbreviated name of days in Spanish.}
\format{
A vector with 7 elements.
}
\usage{
dia_abr
}
\description{
A character vector containing the abbreviated name of the 7 days of the week in Spanish.
Spanish diacritics are kept. If you need names without diacritics, see
\link[dendrohist]{dia_abr2} also in this package.
}
\author{
R. Molowny-Horas
}
\keyword{date}
