% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dia_abr2.R
\docType{data}
\name{dia_abr2}
\alias{dia_abr2}
\title{Abbreviated name of days in Spanish.}
\format{
A vector with 7 elements.
}
\usage{
dia_abr2
}
\description{
A character vector containing the abbreviated name of the 7 days of the week in Spanish.
Spanish diacritics have been removed. If you need names with diacritics, see
\link[dendrohist]{dia_abr} also in this package.
}
\author{
R. Molowny-Horas
}
\keyword{date}
