% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dia_nombre.R
\docType{data}
\name{dia_nombre}
\alias{dia_nombre}
\title{Name of days in Spanish.}
\format{
A vector with 7 elements.
}
\usage{
dia_nombre
}
\description{
A character vector containing the name of the 7 days of the week in Spanish.
Spanish diacritics are kept. If you need names without diacritics, see
\link[dendrohist]{dia_nombre2} also in this package.
}
\author{
R. Molowny-Horas
}
\keyword{date}
