% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ggmmss - copia.R
\name{extract_ggmmss}
\alias{extract_ggmmss}
\title{Convert GGMMSS coordinate to decimal number.}
\usage{
extract_ggmmss(x, test_range = T)
}
\arguments{
\item{x}{character, it must be in the format GGGMMSS. A minus or plus sign is allowed, i.e. +GGGMMSS or
-GGGMMSS are accepted (or +/-MMSS, +/-MSS, +/-SS, +/-S).}

\item{test_range}{logical, if set to TRUE a warning
message is issued if the range of the coordinates (either seconds, minutes or degrees) are outside reasonable values.
See Details below.}
}
\value{
The coordinate as a decimal number.
}
\description{
\code{extract_ggmmss} extracts the degrees, minutes and seconds from the input GGMMSS and converts it
to a decimal number.
}
\details{
The coordinates in the input \code{x} must be GGGMMSS, +GGGMMSS or -GGGMMSS.
The allowed range for seconds and minutes is \link{-60, 60}.
However, we allow for degrees to be in the range \link{-360, 360}, not the usual \link{-180, 180} for longitude or
\link{-90, 90} for latitude.
}
\examples{
# Result is 34.36556
extract_ggmmss("342156")

# Result is -434.3656 with a warning.
extract_ggmmss("-4342156")

# Same without warning.
extract_ggmmss("-4342156", F)
}
