% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood_variogram.R
\name{flood_variogram}
\alias{flood_variogram}
\title{Title}
\usage{
flood_variogram(
  x,
  formula = log(num_events) ~ 1,
  selection = "all",
  cutoff = 200,
  width = 1,
  model_type = "Sph",
  nboot = 100,
  PR = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{nboot}{}
}
\description{
Title
}
