% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap_year.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Check for leap year}
\usage{
is_leap_year(year)
}
\arguments{
\item{year}{a \code{Date} or \code{numeric} vector.}
}
\value{
\code{logical} vector of the same length as the input informing whether or not each element
corresponds to a leap year.
}
\description{
\code{is_leap_year} checks input (which can be \code{numeric} or \code{Date}) for leap year.
}
\details{
First version prepared with Google Gemini. Minor changes afterwards.
}
\examples{
x <- c("02/27/37", "02/27/46", "01/14/92", "02/28/96", "02/01/99")
is_leap_year(as.Date(x, "\%m/\%d/\%y"))
}
