% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_flood_DB.R
\name{merge_flood_DB}
\alias{merge_flood_DB}
\title{Title}
\usage{
merge_flood_DB(...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{

# Correct data.frames.
df1 <- data.frame(x = runif(10), y = sample(letters[1:10]), z = sample(c(F, T), 10, replace = T))
df2<- data.frame(x = runif(10), y = sample(letters[1:10]), z = sample(c(F, T), 10, replace = T))
df3 <- data.frame(x = runif(10), y = sample(letters[1:10]), z = sample(c(F, T), 10, replace = T))

}
