% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mes_abr.R
\docType{data}
\name{mes_abr}
\alias{mes_abr}
\title{Name of months in Spanish, abbreviated to three characters.}
\format{
A vector with 12 elements.
}
\usage{
mes_abr
}
\description{
A character vector containing the abbreviated name of the 12 months of the year in Spanish.
}
\author{
R. Molowny-Horas
}
\keyword{date}
