% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mes_nombre.R
\docType{data}
\name{mes_nombre}
\alias{mes_nombre}
\title{Name of months in Spanish}
\format{
A vector with 12 elements.
}
\usage{
mes_nombre
}
\description{
A character vector containing the full name of the 12 months of the year in Spanish.
}
\author{
R. Molowny-Horas
}
\keyword{date}
