% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_to_number.R
\name{month_to_number}
\alias{month_to_number}
\title{Convert English name of month to number}
\usage{
month_to_number(x, to_lower = T, abbreviated = T)
}
\arguments{
\item{x}{string vector containing the names of months.}

\item{abbreviated}{logical, if set to TRUE \code{month_to_number} will additionally
look for abbreviated names (e.g. "Jan" instead of "January"; see \code{month.abb} for
accepted abbreviations).}

\item{tolower}{logical, if set to TRUE all names are translated to lower case.}
}
\value{
A numeric vector of the same length as 'x'.
}
\description{
\code{month_to_number} converts a vector of English month names into numbers, i.e.
'January' will be 1, 'February' will be 2 and so on.
}
\details{
Simple implementation of \code{match}.
}
\examples{
x <- c("Asd","October", "January", "Dec")
print(month_to_number(x))
}
