% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munic_CNIG.R
\docType{data}
\name{munic_CNIG}
\alias{munic_CNIG}
\title{Spanish municipal names and longitude-latitude coordinates.}
\format{
A data frame.
}
\source{
Go to web site of the Centro Nacional de Información Geográfica
(https://centrodedescargas.cnig.es/CentroDescargas/index.jsp), click on
"Información geográfica de referencia" and on
"Nomenclátor Geográfico de Municipios y Entidades de Población".
}
\usage{
munic_CNIG
}
\description{
A dataset all Spanish municipal names and other information, as well as their
longitude-latitude coordinates.
}
\author{
R. Molowny-Horas
}
\keyword{census,}
\keyword{demography.}
