% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munic_INE.R
\docType{data}
\name{munic_INE}
\alias{munic_INE}
\title{Spanish municipal names and codes}
\format{
A data frame with 5 rows and 2 variables:
}
\source{
Downloaded from the Spanish Instituto Nacional de Estadística
}
\usage{
munic_INE
}
\description{
A dataset all Spanish municipal names and codes, as well as their corresponding autonomous
communities codes.
}
\author{
R. Molowny-Horas
}
\keyword{census,}
\keyword{demography}
