% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmatch.R
\name{rmatch}
\alias{rmatch}
\title{Pattern/Value matching}
\usage{
rmatch(pattern, text, regexpr = F)
}
\arguments{
\item{pattern}{character vector of regular expressions to be matched in 'text'.}

\item{text}{a character vector where matches are sought.}

\item{regexpr}{logical, if set to FALSE (default) \link[base]{match} will be used; otherwise,
\link[base]{regexpr} will be implemented (this choice implies that partial matches are also allowed).}
}
\value{
A \code{list} object with the same length as \code{pattern}. Each element of the output consists of a
vector with the indices of the matches of the corresponding element of 'pattern' in vector 'text'.
}
\description{
\code{rmatch} returns a list of matches of its first argument in its second.
}
\details{
Simple implementation of \link[base]{regexpr} or \link[base]{match}.
}
\examples{
x <- c("NA", "asdf", "aaa", "ppNA")
y <- c("ASDF", "asdf", "ghhjjh", "aaa", "sdgfdgfhj", "NA", "aaa", "Pasdf")

# Notice the differences.
print(rmatch(x, y))
print(rmatch(x, y, T))

print(rmatch(y, x))
print(rmatch(y, x, T))

}
