% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Shift a vector.}
\usage{
roll(x, n)
}
\arguments{
\item{x}{a vector of length > 1.}

\item{n}{an integer stating the amount by which \code{x} will be shifted or rolled.}
}
\value{
A shifted vector of the same length as \code{x}.
}
\description{
\code{roll} shifts a vector right or left with a wrap-up effect.
}
\details{
Found at \code{https://stackoverflow.com/questions/26997586/shifting-a-vector}.
A positive \code{n} moves the vector to the right (as if "pushing" from the left)
so that elements at the right tip of the vector wrap up from the left.
}
\examples{
print(roll(letters[1:5], 2))
}
