% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slash_split.R
\name{slash_split}
\alias{slash_split}
\title{Split string at slash position}
\usage{
slash_split(x, trim = T)
}
\arguments{
\item{x}{character vector.}

\item{trim}{logical, if set to TRUE all leading and trailing whitespaces are
removed \emph{after} the split.}
}
\value{
A \code{data.frame} with two named columns "left" and "right", containing
the left and right side of the string that has been splitted at the slash
location. If no slash is present, "left" will be equal to the original
string and "right" will be NA.

If the slash is located at the first character, "left" will be an empty
character and "right" will contain the string at the right side of the
slash. The opposite will happen if the slash is at the end of the string.

The number of roes of the output \code{data.frame} is equal to the length
of vector \code{x}.
}
\description{
\code{slash_split} splits a string into two at the location of the first slash "/".
}
\details{
The implementation does not take into account more than one slash.
If there is none, no splitting is applied.
}
\examples{
x <- c("This is left/ and right", "/ Slash at start",
"Slash at the end /", "No slash at all")
print(slash_split(x))
}
