% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_imputations.R
\name{ts_imputations}
\alias{ts_imputations}
\title{Imputation of NA in time series}
\usage{
ts_imputations(x, methods = NULL, verbose = T, ...)
}
\arguments{
\item{x}{a \code{numeric} vector or \code{ts} object containing the time series.
Internally it is coerced into a vector.}

\item{methods}{\code{character} vector containing the names of the imputation methods to
be applied.}

\item{verbose}{\code{logical}, if set to TRUE messages are printed on the screen.}

\item{...}{Other input parameters for some of the imputation functions.}
}
\value{
A vector with the reconstructed time series.
}
\description{
It inputs missing values (labelled as NA) in a time series with several
methodologies.
}
\details{
Additional details...
}
\examples{
data("EuStockMarkets")
y <- as.vector(EuStockMarkets[,1]) # Index DAX

# Introduce 20\% random NAs.
p <- 0.2
yNA <- y
yNA[sample(length(y), round(length(y)*p))] <- NA
yrepaired <- ts_imputations(yNA, method = "kalman")
plot(y, type = "l")
points(yrepaired[[1]], type = "l", col = "red")

}
