% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cedex_basin.R
\name{read_cedex_basin}
\alias{read_cedex_basin}
\title{Read Spanish basin data from the \emph{cedex} site.}
\usage{
read_cedex_basin(table = NULL, str_url = NULL, English = F)
}
\arguments{
\item{table}{\code{character} with the name of the file to retrieve from the
\emph{cedex} site, without extension. If not given, the default value is "afliq".}

\item{str_url}{string with the \code{url} of the \emph{cedex} site. Value by default is
"https://ceh-flumen64.cedex.es/anuarioaforos//anuario-2019-2020/EBRO/", corresponding to the
Ebro basin.
Notice that, internally, \code{read_cedex_basin} will paste \code{table} and \code{str_url} together,
so any missing slash at the end of \code{str_url} will raise an error.}

\item{English}{logical, if set to TRUE the names of the columns of the output \code{data.frame} are
translated into English. If set to FALSE (default), Spanish names (as in the original files) are kept.}
}
\value{
A \code{data.frame}. Column labels vary depending on the value of "table":
\itemize{
\item \code{table = "afliq"}: in Spanish, "indroea", "fecha", "altura" and "caudal"; in English, "Station", "Date", "Height" and "Flow".
\item \code{table = "mensual_a"}: in Spanish, "indroea", "anomes", "hmedmes", "qmedmes", "apormedmes", "qcmes", #' "hcmes", "dia_cmes", "qnmes", "hnmes", "dia_nmes", "orig_dato_id";
in English, "Station", "Date", "Height", "Flow", "Streamflow", "Flow_max", "Height_max", "Flow_max_day", "Flow_min", "Height_min", "Flow_min_day", "Orig_data").
}
}
\description{
\code{read_cedex_basin} retrieves time series datasets and statistics for
several Spanish basins from the Centro de Estudios y Experimentación
de Obras Públicas (CEDEX).
}
\details{
To see a description of the files to retrieve see
"https://ceh.cedex.es/anuarioaforos/demarcaciones.asp" and go to the basin you want
to get data for.
}
\examples{
x <- read_cedex_basin()
}
